Rename DataDefinition namespace to Definition
