+++ Robust.Shared/Serialization/Manager/Definition/DataDefinition.Delegates.cs
@@ -1,7 +1,7 @@
 ﻿using Robust.Shared.Serialization.Manager.Result;
 using Robust.Shared.Serialization.Markdown.Mapping;
 
-namespace Robust.Shared.Serialization.Manager.DataDefinition
+namespace Robust.Shared.Serialization.Manager.Definition
 {
     public partial class DataDefinition
     {
